requirements.txt
gevent~=21.12.0
requests~=2.27.1
pyyaml~=6.0
bs4~=0.0.1
beautifulsoup4~=4.11.1
loguru~=0.6.0